Move validation tests to overture-schema-validation
Validation test files (type_analysis, error_formatting, data_display,
heterogeneous_collections) move to the validation package. Validate
command and function tests extracted from CLI test modules.

CLI retains list-types, json-schema, and plugin loading tests.
