+++ packages/overture-schema-validation/tests/test_data_display.py
@@ -3,7 +3,7 @@
 from io import StringIO
 
 from conftest import build_feature
-from overture.schema.cli.data_display import (
+from overture.schema.validation.data_display import (
     create_feature_display,
     extract_feature_data,
     format_field_value,
