+++ packages/overture-schema-validation/tests/test_error_formatting.py
@@ -7,12 +7,12 @@ from unittest.mock import patch
 import pytest
 from click.testing import CliRunner
 from overture.schema.cli.commands import cli
-from overture.schema.cli.error_formatting import (
+from overture.schema.validation.error_formatting import (
     format_path,
     group_errors_by_discriminator,
     select_most_likely_errors,
 )
-from overture.schema.cli.type_analysis import introspect_union
+from overture.schema.validation.type_analysis import introspect_union
 from pydantic import BaseModel, Field, TypeAdapter, ValidationError
 from rich.console import Console
 
@@ -134,7 +134,7 @@ properties:
         buffer = StringIO()
         captured_console = Console(file=buffer, force_terminal=False)
 
-        with patch("overture.schema.cli.commands.stderr", captured_console):
+        with patch("overture.schema.validation.commands.stderr", captured_console):
             result = cli_runner.invoke(cli, ["validate", filename])
 
         assert result.exit_code == 1
@@ -181,7 +181,7 @@ properties:
         buffer = StringIO()
         captured_console = Console(file=buffer, force_terminal=False)
 
-        with patch("overture.schema.cli.commands.stderr", captured_console):
+        with patch("overture.schema.validation.commands.stderr", captured_console):
             result = cli_runner.invoke(cli, ["validate", filename])
 
         assert result.exit_code == 1
