+++ packages/overture-schema-validation/tests/test_heterogeneous_collections.py
